Add jdoc and refactor some classes
