+++ backend/src/main/java/edu/kit/quak/Savable.java
@@ -3,6 +3,11 @@ package edu.kit.quak;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 
+/**
+ * Defines the ID of every savable entity.
+ *
+ * @author Henrik K
+ */
 @Entity
 public interface Savable {
     @Id
